PHPUnit 3.4.x mocks concrete methods too.
When calling getMockForAbstractClass() on an abstract class that implements
an interface, PHPUnit 3.4.x will mock abstract methods (those method that
are part of the interface and not implemented in the abstract class), but it
seems it will also mock concrete methods (those that appear in the interface
but have an implementation in the abstract class).
Special care must be taken when dealing with such classes.

This helps fix an issue with several modules where the tests would just fail
on the PHP 5.2 builder in our CI process.